Rename subsystem_rest_resolver.rs to just resolver.rs for consistency
